--- src/contradictory_claims/claims/__init__.py
@@ -1,9 +1,0 @@
-# -*- coding: utf-8 -*-
-
-"""Code for Identifying core claims from a paragraph."""
-
-from .extract_claims import *  # noqa:F401,F403
-from .models import *  # noqa:F401,F403
-from .dataset_reader import *  # noqa:F401,F403
-from .predictors import *  # noqa:F401,F403
-from .utils import *  # noqa:F401,F403
